Rework file list generation in copy-dotfiles
Previously, copy-dotfiles would simply copy everything that appeared in
the top level of the repository, which includes the contents of
directories for which there are any untracked files. This is desirable
for .vim/ but undesirable for other directories.

This change reworks copy-dotfiles to only copy the files tracked in the
repository, with a special case for .vim/ (the full directory contents
are copied).