Remap :W to ,W in vim to avoid expanding "W" on the command line
The previous :W mapping worked well, but clashed with any capital W
typed on the vim command line. For example, searching for something
starting with W (typing "/W", etc.) would replace the W to the full
command line mapping for :W. This effectively prevented the use of W on
the vim command line. Changing this to a normal mode mapping avoids this
issue.