Use vim-tmux-navigator for intelligent vim and tmux navigation
This allows Ctrl+[HJKL] to be used to traverse both vim splits and tmux
panes. If the active tmux pane is running vim, vim switches splits when
one of the navigation keys is pressed, or forwards the command back to
tmux if the furthest split in the requested direction was already
focused.

This uses a modified version of the tmux configuration recommended by
vim-tmux-navigator for Ctrl+L. If the active tmux window only has a
single pane, then Ctrl+L is forwarded onto the terminal in order to
clear the screen. For tmux windows with multiple panes, <Prefix>Ctrl+L
can be used instead.

In order for this to work with vim instances running as root, the TMUX
environment variable must be set properly. To ensure this happens with
sudo, the entire environment may be preserved with:

$ sudo -E vim

Alternatively, just the value of TMUX can be preserved with:

$ sudo TMUX="${TMUX}" vim