In vim, use Python version from shebang line if present with Syntastic
If a shebang line containing an explicit Python version number is
present in the current buffer, then Syntastic is configured to call
flake8 with that particular version of Python. In this case, the
secondary "python" checker is also disabled, as there isn't a good way
to override the executable with the specified version of Python.