Add setup-caff utility to configure caff for offline key signing
setup-caff is meant to be run in an offline live environment after the
key(s) to be signed and the signing private key have all been imported
into the current user's keyring. This populates ~/.caffrc and copies all
the public keys from the current user's keyring to caff's signing
keyring.