Add the git cg alias for graph comparisons between branches
The git cg alias shows the log (with --graph) between the current branch
and the specified branch's remote tracking branch. This is useful for
comparing the log on a feature branch that doesn't track the branch it
will be merged back into.

For example, running "git cg master" on a feature branch will show the
log (with --graph) of commits on the feature branch that are not in
master's remote tracking branch (e.g. origin/master). Merging the
current branch back into master will bring in all the commits listed by
the aforementioned git cg command.

git cg is similar to the existing git ux and git ul aliases, except that
a different comparison branch may be specified with git cg. git ux and
git ul only show the differences between the current branch and its own
remote tracking branch.