Skip dotfiles repository updates that have already been applied
This prevents dotfiles-auto-update from being executed if the branch was
updated between the time the repository was checked and when the shell
tries to actually fetch the update. This can occur when an update check
is triggered (typically when resuming or opening a new shell session)
and "git pull" is run manually before __auto_update_check in .bashrc is
actually invoked to pull the update.