Reload env vars from tmux if the tmux socket change time has changed
tmux sessions store environment variable values, such as DISPLAY and
SSH_AUTH_SOCK, which can be updated by tmux clients (e.g. when sessions
are detached or attached).

Similar to re-sourcing .bashrc if any of the dotfiles have changed, this
change updates environment variable values from tmux if the shell is
within a tmux session and the tmux session socket's change time has
changed. Based on empirical observation, the socket's change time
appears to change when sessions are detached or attached, and possibly
under other circumstances. This new feature provides an automatic method
for variables such as SSH_AUTH_SOCK to remain up to date when
reattaching to a tmux session.